refactor: 정렬 방식 표기 수정
