+++ app/src/main/res/layout/fragment_home.xml
@@ -5,6 +5,9 @@
 
     <data>
 
+        <variable
+            name="vm"
+            type="com.ddd4.synesthesia.beer.presentation.ui.home.viewmodel.HomeViewModel" />
     </data>
 
     <androidx.core.widget.NestedScrollView
@@ -45,7 +48,6 @@
                     android:layout_height="wrap_content"
                     android:layout_margin="12dp"
                     android:fontFamily="@font/notosans_kr_regular"
-                    android:text="기본 정렬"
                     android:textColor="@color/white"
                     android:textSize="12sp"
                     android:drawableEnd="@drawable/ic_arrow_drop_down"
@@ -53,7 +55,10 @@
                     android:gravity="center"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
+                    app:layout_constraintTop_toTopOf="parent"
+                    app:sortTypeText="@{vm.sortType}"
+
+                    />
 
             </androidx.constraintlayout.widget.ConstraintLayout>
