+++ app/src/main/res/values/strings.xml
@@ -19,4 +19,8 @@
     <string name="rate">%.0f점</string>
     <string name="back_press">\'뒤로\' 버튼을 한번 더 누르시면 종료됩니다</string>
     <string name="profile_name">%s님</string>
+
+    <string name="sort_default">기본 정렬</string>
+    <string name="sort_comment">후기 정렬</string>
+    <string name="sort_review">리뷰 정렬</string>
 </resources>
